[#3236]Move all Message classes to Akvo namespace.
