+++ GAE/war/WEB-INF/rest-servlet.xml
@@ -30,7 +30,7 @@
     <bean name="/data_point_assignments" class="org.akvo.flow.rest.DataPointAssignmentRestService" lazy-init="true" />
     <bean name="/device_groups" class="org.waterforpeople.mapping.app.web.rest.DeviceGroupRestService" lazy-init="true" />
     <bean name="/devices" class="org.waterforpeople.mapping.app.web.rest.DeviceRestService" lazy-init="true" />
-    <bean name="/messages" class="org.waterforpeople.mapping.app.web.rest.MessageRestService" lazy-init="true" />
+    <bean name="/messages" class="org.akvo.flow.rest.MessageRestService" lazy-init="true" />
     <bean name="/notification_subscriptions" class="org.waterforpeople.mapping.app.web.rest.NotificationSubscriptionRestService" lazy-init="true" />
     <bean name="/placemarks" class="org.waterforpeople.mapping.app.web.rest.PlacemarkRestService" lazy-init="true" />
     <bean name="/question_answers" class="org.waterforpeople.mapping.app.web.rest.QuestionAnswerRestService" lazy-init="true" />
